+++ src/featherlight.gallery.js
@@ -41,15 +41,17 @@
 					var fl = this,
 						$img = fl.$instance.find('img');
 
+					fl.$currentTarget = $(event.currentTarget);
+
 					$img.load(function(){
 						$img.stop().fadeTo(fl.config.gallery.fadeIn,1);
 					});
 
 					fl.$instance.on('next.'+fl.config.namespace+' previous.'+fl.config.namespace, function(event){
 							var offset = event.type === 'next' ? +1 : -1;
-							var $nx = $gallery.eq(($gallery.index(fl.$elm)+offset) % $gallery.length);
+							var $nx = $gallery.eq(($gallery.index(fl.$currentTarget)+offset) % $gallery.length);
 							$img.fadeTo(fl.config.gallery.fadeOut,0.2);
-							fl.$elm = $nx;
+							fl.$currentTarget = $nx;
 							$img[0].src = $nx.attr('href');
 						});